Rename: Chat -> ChatRoom 파일명 변경
